Move function tests inside `src` folder.
